src/storage/stats-daemon/requirements.txt
bitarray
bottle
gevent
geventhttpclient
psutil
sandesh_common
sandesh
discoveryclient
